Limit tox to test_*.py files; report warnings
Tox was issuing warnings because of support classes within the test
directory that are named with "Test*" patterns, but are not TestCase
subclasses (``kvstore.TestKVStore`` and ``storage.TestStorage``).

Fix this by limiting test discovery to files that match the
``test_*.py`` pattern. Also, add the "-rw" switch to the py.test
invocation to report warnings.